pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'


[project]
name='noisetools'
version='0.1.0-dev0'
description='A Python toolbox with all the specialty signal analysis tools for my PhD.'
readme='README.md'
authors=[
    {'name' = 'Josephine Siebert Pockelé', email='j.s.pockele@tudelft.nl'}
]
requires-python='>=3.9'
dependencies=[
    'numpy>=1.26.4',
    'scipy>=1.12.0',
    'matlabengine>=23.2.2'
]
classifiers=[
    'Development Status :: 1 - Planning',
    'Intended Audience :: Education',
    'Intended Audience :: Science/Research',
    'Natural Language :: English',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: Microsoft :: Windows :: Windows 10',
    'Operating System :: Microsoft :: Windows :: Windows 11',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Topic :: Scientific/Engineering',
    'Topic :: Scientific/Engineering :: Physics',
    'Topic :: Utilities'
]

[tool.setuptools]
include-package-data = true

[project.optional-dependencies]
dev = [
    'pytest>=8.0.2',
    'twine>=5.0.0'
]

[project.urls]
source = 'https://gitlab.tudelft.nl/jpockele/josephines-bag-of-tricks'
